requirements.txt
Flask==3.0.0
requests==2.31.0
python-dotenv==1.0.0
groq==0.4.1
google-generativeai==0.3.2
moviepy==1.0.3
Pillow==10.1.0
gTTS==2.5.0
beautifulsoup4==4.12.2